backend/requirements.txt
anyio==4.0.0
cassandra-driver==3.27.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
exceptiongroup==1.1.3
fastapi==0.95.1
geomet==0.2.1.post1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.3
passlib==1.7.4
pycparser==2.21
pydantic==1.10.13
PyJWT==2.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.8.0
uvicorn==0.21.1
uvloop==0.17.0
